feature: CRUD

+++ src/dto/UserDTO.java
@@ -5,7 +5,7 @@ public class UserDTO {
     private String user_pw;
     private String name;
     private ROLE role;
-    private long class_id;
+    private int class_id;
 
     public String getUser_id() {
         return user_id;
@@ -23,11 +23,11 @@ public class UserDTO {
         return role;
     }
 
-    public long getClass_id() {
+    public int getClass_id() {
         return class_id;
     }
 
-    public UserDTO(String user_id, String user_pw, String name, ROLE role, long class_id) {
+    public UserDTO(String user_id, String user_pw, String name, ROLE role, int class_id) {
         this.user_id = user_id;
         this.user_pw = user_pw;
         this.name = name;
